A few more places where large images cause int overflows...

See merge request RD/idg!77

+++ idg-lib/src/CPU/common/InstanceCPU.h
@@ -148,7 +148,7 @@ class InstanceCPU : public KernelsInstance {
    * @param subgrid_size size of the subgrids
    * @return The number of wtiles
    */
-  virtual int init_wtiles(int grid_size, int subgrid_size);
+  virtual size_t init_wtiles(size_t grid_size, int subgrid_size);
 
  protected:
   void compile(Compiler compiler, Compilerflags flags);
